修复 一些导致闪退与加载失败的问题 修复 注销后重新登陆时数据加载问题 优化 本地数据加载线程并发

+++ app/src/main/java/tool/xfy9326/naucourse/Methods/ImageMethod.java
@@ -54,26 +54,30 @@ public class ImageMethod {
             if (responseBody != null) {
                 InputStream inputStream = responseBody.byteStream();
                 Bitmap bitmap = BitmapFactory.decodeStream(inputStream);
+                response.close();
                 return saveBitmap(bitmap, downloadPath);
             }
         }
+        response.close();
         return false;
     }
 
     public static boolean saveBitmap(Bitmap bitmap, String path) throws IOException {
-        File file = new File(path);
-        if (file.exists()) {
-            //noinspection ResultOfMethodCallIgnored
-            file.delete();
-        }
-        if (IO.createPath(file)) {
-            if (file.createNewFile()) {
-                FileOutputStream fileOutputStream = new FileOutputStream(file);
-                bitmap.compress(Bitmap.CompressFormat.JPEG, 100, fileOutputStream);
-                fileOutputStream.flush();
-                fileOutputStream.close();
-                bitmap.recycle();
-                return true;
+        if (bitmap != null) {
+            File file = new File(path);
+            if (file.exists()) {
+                //noinspection ResultOfMethodCallIgnored
+                file.delete();
+            }
+            if (IO.createPath(file)) {
+                if (file.createNewFile()) {
+                    FileOutputStream fileOutputStream = new FileOutputStream(file);
+                    bitmap.compress(Bitmap.CompressFormat.JPEG, 100, fileOutputStream);
+                    fileOutputStream.flush();
+                    fileOutputStream.close();
+                    bitmap.recycle();
+                    return true;
+                }
             }
         }
         return false;
